requirements.txt
aiogram==3.20.0
python-dotenv==1.0.0
remnawave_api
psutil==5.9.6
